remove update logic and simplify extension.ts
