+++ jdtls.ext/com.microsoft.java.maven.plugin/src/main/java/com/microsoft/java/maven/AddDependencyHandler.java
@@ -68,31 +68,24 @@ public class AddDependencyHandler {
     }
 
     private static WorkspaceEdit importEdit(AddDependencyParams params, 
-            ICompilationUnit unit, CompilationUnit astRoot) {
+            ICompilationUnit unit, CompilationUnit astRoot) throws CoreException {
         // import the new class
         final CompilationUnitChange cuChange = new CompilationUnitChange("", unit);
         final ImportRewrite importRewrite = CodeStyleConfiguration.createImportRewrite(astRoot, true);
         importRewrite.addImport(params.fullClassName);
-        try {
-            cuChange.setEdit(importRewrite.rewriteImports(null));
-            return ChangeUtil.convertToWorkspaceEdit(cuChange);
-        } catch (CoreException e) {
-            return new WorkspaceEdit();
-        }
+        cuChange.setEdit(importRewrite.rewriteImports(null));
+        return ChangeUtil.convertToWorkspaceEdit(cuChange);
     }
 
-    private static WorkspaceEdit replaceEdit(AddDependencyParams params, ICompilationUnit unit) {
+    private static WorkspaceEdit replaceEdit(AddDependencyParams params, 
+            ICompilationUnit unit) throws JavaModelException {
         // replace old class name with new one
-        try {
-            final Map<String, List<TextEdit>> textEdits = new HashMap<>();
-            textEdits.put(unit.getCorrespondingResource().getLocationURI().toString(),
+        final Map<String, List<TextEdit>> textEdits = new HashMap<>();
+        textEdits.put(unit.getCorrespondingResource().getLocationURI().toString(),
                     Arrays.asList(new TextEdit(new Range(new Position(params.line, params.character), 
                     new Position(params.line, params.character + params.length)), 
                     params.fullClassName.substring(params.fullClassName.lastIndexOf('.') + 1))));
-            return new WorkspaceEdit(textEdits);
-        } catch (JavaModelException e) {
-            return new WorkspaceEdit();
-        }
+        return new WorkspaceEdit(textEdits);
     }
 
     private static PosInfo getPosInfo(String pomPath, String targetDependency) {
@@ -108,60 +101,58 @@ public class AddDependencyHandler {
             posInfo = posHandler.getPosInfo();
             return posInfo;
         } catch (SAXException | IOException | ParserConfigurationException e) {
+            // there is no pom.xml, or fail to get parser
             return null;
         }
     }
 
-    private static WorkspaceEdit pomEdit(AddDependencyParams params, ICompilationUnit unit) {
-        IPath path = null;
-        try {
-            path = unit.getJavaProject().getCorrespondingResource().getLocation().append("pom.xml");
-            final String info[] = params.artifactInfo.replaceAll(" ", "").split(":");
-            if (info.length < 3) {
-                return new WorkspaceEdit();
-            }
-            final PosInfo posInfo = getPosInfo(path.toString(), info[0] + ":" + info[1]);
-            final String linesep = System.lineSeparator(); 
-            String newtext = "";
-            final String pomUriString = unit.getJavaProject().getCorrespondingResource()
-                    .getLocationURI().toString() + "/pom.xml";
-            final Map<String, List<TextEdit>> textEdits = new HashMap<>();
+    private static WorkspaceEdit pomEdit(AddDependencyParams params, 
+            ICompilationUnit unit) throws JavaModelException {
+        IPath path = unit.getJavaProject().getCorrespondingResource().getLocation().append("pom.xml");
+        final String info[] = params.artifactInfo.replaceAll(" ", "").split(":");
+        if (info.length < 3) {
+            return new WorkspaceEdit();
+        }
+        final PosInfo posInfo = getPosInfo(path.toString(), info[0] + ":" + info[1]);
+        final String linesep = System.lineSeparator(); 
+        String newtext = "";
+        final String pomUriString = unit.getJavaProject().getCorrespondingResource()
+                .getLocationURI().toString() + "/pom.xml";
+        final Map<String, List<TextEdit>> textEdits = new HashMap<>();
 
-            if (posInfo != null) {
-                if (posInfo.needAddDependency == false) {
-                    textEdits.put(pomUriString, new ArrayList<TextEdit>());
-                } else if (posInfo.alreadyHasDependencies == false) {
-                    final int space = 2;
-                    newtext = linesep + StringUtils.repeat(" ", space) + "<dependencies>" + linesep + 
-                            StringUtils.repeat(" ", space + 2) + "<dependency>" + linesep + 
-                            StringUtils.repeat(" ", space + 4) + "<groupId>" + info[0] + "</groupId>" + 
-                            linesep + StringUtils.repeat(" ", space + 4) + "<artifactId>" + info[1] + 
-                            "</artifactId>" + linesep + StringUtils.repeat(" ", space + 4) + "<version>" + 
-                            info[2] + "</version>" + linesep + StringUtils.repeat(" ", space + 2) + "</dependency>" + 
-                            linesep + StringUtils.repeat(" ", space) + "</dependencies>" + linesep;
-                    textEdits.put(pomUriString, 
-                        Arrays.asList(new TextEdit(new Range(posInfo.pos, posInfo.pos), newtext)));
-                } else {
-                    final int space = posInfo.pos.getCharacter();
-                    newtext = linesep + StringUtils.repeat(" ", space + 2) + "<dependency>" + linesep + 
-                        StringUtils.repeat(" ", space + 4) + "<groupId>" + info[0] + "</groupId>" + linesep + 
-                        StringUtils.repeat(" ", space + 4) + "<artifactId>" + info[1] + "</artifactId>" + 
-                        linesep + StringUtils.repeat(" ", space + 4) + "<version>" + info[2] + "</version>" + 
-                        linesep + StringUtils.repeat(" ", space + 2) + "</dependency>" + linesep + 
-                        StringUtils.repeat(" ", space);
-                    textEdits.put(pomUriString, 
-                        Arrays.asList(new TextEdit(new Range(posInfo.pos, posInfo.pos), newtext)));
-                }
-                return new WorkspaceEdit(textEdits);
+        if (posInfo != null) {
+            if (posInfo.needAddDependency == false) {
+                textEdits.put(pomUriString, new ArrayList<TextEdit>());
+            } else if (posInfo.alreadyHasDependencies == false) {
+                final int space = 2;
+                newtext = linesep + StringUtils.repeat(" ", space) + "<dependencies>" + linesep + 
+                        StringUtils.repeat(" ", space + 2) + "<dependency>" + linesep + 
+                        StringUtils.repeat(" ", space + 4) + "<groupId>" + info[0] + "</groupId>" + 
+                        linesep + StringUtils.repeat(" ", space + 4) + "<artifactId>" + info[1] + 
+                        "</artifactId>" + linesep + StringUtils.repeat(" ", space + 4) + "<version>" + 
+                        info[2] + "</version>" + linesep + StringUtils.repeat(" ", space + 2) + "</dependency>" + 
+                        linesep + StringUtils.repeat(" ", space) + "</dependencies>" + linesep;
+                textEdits.put(pomUriString, 
+                    Arrays.asList(new TextEdit(new Range(posInfo.pos, posInfo.pos), newtext)));
             } else {
-                return new WorkspaceEdit();
+                final int space = posInfo.pos.getCharacter();
+                newtext = linesep + StringUtils.repeat(" ", space + 2) + "<dependency>" + linesep + 
+                    StringUtils.repeat(" ", space + 4) + "<groupId>" + info[0] + "</groupId>" + linesep + 
+                    StringUtils.repeat(" ", space + 4) + "<artifactId>" + info[1] + "</artifactId>" + 
+                    linesep + StringUtils.repeat(" ", space + 4) + "<version>" + info[2] + "</version>" + 
+                    linesep + StringUtils.repeat(" ", space + 2) + "</dependency>" + linesep + 
+                    StringUtils.repeat(" ", space);
+                textEdits.put(pomUriString, 
+                    Arrays.asList(new TextEdit(new Range(posInfo.pos, posInfo.pos), newtext)));
             }
-        } catch (JavaModelException e) {
+            return new WorkspaceEdit(textEdits);
+        } else {
             return new WorkspaceEdit();
         }
     }
 
-    public static List<WorkspaceEdit> addDependency(AddDependencyParams params, IProgressMonitor monitor) {
+    public static List<WorkspaceEdit> addDependency(AddDependencyParams params, 
+            IProgressMonitor monitor) throws CoreException, JavaModelException {
         final ICompilationUnit unit = JDTUtils.resolveCompilationUnit(params.uri);
         final CompilationUnit astRoot = ASTResolving.createQuickFixAST(unit, null);
         return Arrays.asList(replaceEdit(params, unit), importEdit(params, unit, astRoot),
